# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentFlow Runtime API
  description: |
    OpenAPI contract for endpoints currently exposed by `cmd/agentflow` runtime.

    Notes:
    - API key authentication is accepted via `X-API-Key` header.
    - Query-string `api_key` is disabled by default for security reasons.
  version: 1.3.0
  contact:
    name: AgentFlow Team
    url: https://github.com/BaSui01/agentflow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: HTTP API server
...
paths:
...
  /api/v1/agents:
    get:
      tags: [Agent]
      summary: List agents
      operationId: listAgents
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AgentInfo'
        '500':
          description: Internal error
  /api/v1/agents/execute:
    post:
      tags: [Agent]
      summary: Execute agent
      operationId: executeAgent
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecuteRequest'
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AgentExecuteResponse'
        '400':
          description: Invalid request
        '404':
          description: Agent not found
        '500':
          description: Execution failed
  /api/v1/agents/execute/stream:
    post:
      tags: [Agent]
      summary: Stream agent execution (SSE)
      operationId: executeAgentStream
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecuteRequest'
      responses:
        '200':
          description: SSE stream of agent execution events (token, tool_call, tool_result, error, [DONE])
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request
        '404':
          description: Agent not found
        '500':
          description: Execution failed
  /api/v1/agents/plan:
    post:
      tags: [Agent]
      summary: Plan agent execution
      operationId: planAgent
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentExecuteRequest'
      responses:
        '200':
          description: Execution plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '400':
          description: Invalid request
        '404':
          description: Agent not found
        '500':
          description: Plan failed
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Envelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          additionalProperties: true
        error:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          description: Unique request identifier for tracing
          example: "req-abc123"
...
    AgentInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        state:
          type: string
        description:
          type: string
        model:
          type: string
        created_at:
          type: string
    AgentExecuteRequest:
      type: object
      required: [agent_id, content]
      properties:
        agent_id:
          type: string
        content:
          type: string
        context:
          type: object
          additionalProperties: true
        variables:
          type: object
          additionalProperties:
            type: string
    AgentExecuteResponse:
      type: object
      properties:
        trace_id:
          type: string
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
        tokens_used:
          type: integer
        cost:
          type: number
        duration:
          type: string
        finish_reason:
          type: string
...